refactor(#656): [Pooja|Vijay] refactor methods to get the localized field Value
