+++ components/VcItem.tsx
@@ -23,10 +23,9 @@ import { Theme } from './ui/styleUtils';
 import { RotatingIcon } from './RotatingIcon';
 import { GlobalContext } from '../shared/GlobalContext';
 import { useTranslation } from 'react-i18next';
-import { LocalizedField } from '../types/vc';
 import { VcItemTags } from './VcItemTags';
 import VerifiedIcon from './VerifiedIcon';
-import i18n, { getVCDetailsForCurrentLanguage } from '../i18n';
+import { getLocalizedField } from '../i18n';
 
 const getDetails = (arg1, arg2, verifiableCredential) => {
   if (arg1 === 'Status') {
@@ -304,18 +303,4 @@ interface VcItemProps {
   onPress?: (vcRef?: ActorRefFrom<typeof vcItemMachine>) => void;
   onShow?: (vcRef?: ActorRefFrom<typeof vcItemMachine>) => void;
   activeTab?: string;
-}
-
-function getLocalizedField(rawField: string | LocalizedField) {
-  if (typeof rawField === 'string') {
-    return rawField;
-  }
-  try {
-    const locales: LocalizedField[] = JSON.parse(JSON.stringify(rawField));
-    const currentLanguage = i18n.language;
-    if (locales.length == 1) return locales[0]?.value;
-    return getVCDetailsForCurrentLanguage(locales, currentLanguage);
-  } catch (e) {
-    return '';
-  }
 }
